Update :- see description (Fantasy Ball)
1. Updated sprint and journal documents.
2. Updated tile Definitions XML
3. Added in test tiles to test collisions.
4. Added in support for generating Maps from images.
5. Updated Level Bounds.
6. Paper calculations for percentage base sections complete.

+++ Fantasy Ball/Code/Game/Ball.cpp
@@ -8,7 +8,7 @@ extern RenderContext* g_theRenderer;
 
 //--------------------------------------------------------------------------------------------------------------------------------------------
 
-Ball::Ball( Game* owner , float health , float cosmeticRadius , float physicsRadius , Vec2 position , Vec2 velocity , eEntityType type /*= BALL */ ) :
+Ball::Ball( Game* owner , int health , float cosmeticRadius , float physicsRadius , Vec2 position , Vec2 velocity , eEntityType type /*= BALL */ ) :
 																								Entity( owner , health , type ),
 																								m_cosmeticRadius( cosmeticRadius ) ,
 																								m_physicsRadius( physicsRadius ) ,
